+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 * `warp_affine` can now be used with `backend='itk'`.
 
+* `kwimage.Points` now contains a "new-v2" style of coco keypoints that it can handle.
+
 ### Fixed
 * Fix issue of coercing detections when segmentations are null.